@@ -6,7 +6,6 @@
     "ng": "ng",
     "build": "webpack --config server/webpack.config.js && ng build",
     "start": "ng serve --proxy-config config/proxy.conf.json",
-    "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "server": "webpack --config server/webpack.config.js && node build/main.js",
@@ -29,37 +28,26 @@
     "core-js": "^2.5.1",
     "express": "^4.15.4",
     "injection-js": "^2.1.0",
-    "mongodb": "^2.2.31",
-    "mongoose": "^4.11.12",
+    "mongodb": "^3.1.1",
+    "mongoose": "^5.2.1",
     "ng2-nouislider": "^1.7.0",
-    "nouislider": "^10.0.0",
+    "nouislider": "^11.1.0",
     "q": "^1.5.0",
     "rxjs": "^5.4.3",
-    "ws": "^3.2.0",
+    "ws": "^5.2.1",
     "zone.js": "^0.8.17"
   },
   "devDependencies": {
     "@angular/cli": "1.4.3",
     "@angular/compiler-cli": "^4.2.4",
     "@angular/language-service": "^4.2.4",
-    "@types/jasmine": "2.6.0",
-    "@types/node": "~8.0.30",
-    "@types/ws": "3.0.2",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node-inspector": "^1.1.1",
+    "@types/webpack-env": "^1.13.6",
+    "@types/ws": "^5.1.2",
+    "codelyzer": "~4.4.2",
     "nodemon": "^1.12.1",
-    "protractor": "~5.1.2",
     "ts-loader": "^2.3.7",
     "ts-node": "~3.3.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.5.2"
+    "tslint": "~5.10.0",
+    "typescript": "~2.9.2"
   }
 }
